Added value features

+++ Lendr.API/Controllers/BorrowersV2Controller.cs
@@ -6,31 +6,30 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Lendr.API.Data;
-using Lendr.API.Models;
 using Lendr.API.Core.Contracts;
 using AutoMapper;
 using System.Collections;
 using Lendr.API.Core.DTO.Borrower;
 using Microsoft.AspNetCore.Authorization;
 using Lendr.API.Core.Exceptions;
-
-using Lendr.API.Core.Models;
+using Microsoft.AspNetCore.OData.Query;
+using Lendr.API.Models;
 
 namespace Lendr.API.Controllers
 {
     //[Route("api/[controller]")]
     [Route("api/v{version:apiVersion}/borrowers")]
-    [ApiVersion("1.0",Deprecated = true)]
+    [ApiVersion("2.0")]
     [ApiController]
     
     [Authorize]
-    public class BorrowersController : ControllerBase
+    public class BorrowersV2Controller : ControllerBase
     {
         private readonly IBorrowerRepository _borrowerRepository;
         private readonly IMapper _mapper;
-        private readonly ILogger<BorrowersController> _logger;
+        private readonly ILogger<BorrowersV2Controller> _logger;
 
-        public BorrowersController(IBorrowerRepository borrowerRepository, IMapper mapper,ILogger<BorrowersController> logger)
+        public BorrowersV2Controller(IBorrowerRepository borrowerRepository, IMapper mapper,ILogger<BorrowersV2Controller> logger)
         {
 
             _borrowerRepository = borrowerRepository;
@@ -38,29 +37,31 @@ namespace Lendr.API.Controllers
             this._logger = logger;
         }
 
-        // GET: api/Borrowers/GetAll
-        [HttpGet("GetAll")]
+        // GET: api/Borrowers
+        [HttpGet]
+        [EnableQuery]
         public async Task<ActionResult<IEnumerable<BorrowerDto>>> GetBorrowers()
         {
-            var borrowers =await _borrowerRepository.GetAllAsync<BorrowerDto>();
-            //var borrowerList = _mapper.Map<IEnumerable<BorrowerDto>>(borrowers);
-            return Ok(borrowers);
-        }
-        // GET: api/Borrowers/?StartIndex=0&pageSize=25&PageNumber=1
-        [HttpGet]
-        public async Task<ActionResult<PagedResult<BorrowerDto>>> GetPagedBorrowers([FromQuery] QueryParameters queryParameters)
-        {
-            var pagedBorrowerResult = await _borrowerRepository.GetAllAsync<BorrowerDto>(queryParameters);
-            //var borrowerList = _mapper.Map<IEnumerable<BorrowerDto>>(borrowers);
-            return Ok(pagedBorrowerResult);
+            var borrowers =await _borrowerRepository.GetAllAsync();
+            var borrowerList = _mapper.Map<IEnumerable<BorrowerDto>>(borrowers);
+            return Ok(borrowerList);
         }
 
         // GET: api/Borrowers/5
         [HttpGet("{id}")]
         public async Task<ActionResult<GetBorrowerDto>> GetBorrower(int id)
         {
+
             var borrower = await _borrowerRepository.GetDetails(id);
-            return Ok(borrower);
+
+            if (borrower == null)
+            {
+                //_logger.LogWarning($"No records found in {nameof(GetBorrower)}");
+                //return NotFound();
+                throw new NotFoundException(nameof(GetBorrower), id);
+            }
+            var borrowerDetails = _mapper.Map<GetBorrowerDto>(borrower);
+            return borrowerDetails;
         }
 
         // PUT: api/Borrowers/5
@@ -73,10 +74,15 @@ namespace Lendr.API.Controllers
             {
                 return BadRequest();
             }
-
+            var resultBorrower = await _borrowerRepository.GetAsync(id);
+            if (resultBorrower == null)
+            {
+                return NotFound();
+            }
+            _mapper.Map(borrower, resultBorrower);
             try
             {
-                await _borrowerRepository.UpdateAsync(id, borrower);
+                await _borrowerRepository.UpdateAsync(resultBorrower);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -97,10 +103,17 @@ namespace Lendr.API.Controllers
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         [Authorize(Roles = "Administrator")]
-        public async Task<ActionResult<BorrowerDto>> PostBorrower(CreateBorrowerDto newBorrower)
+        public async Task<ActionResult<Borrower>> PostBorrower(CreateBorrowerDto newBorrower)
         {
-            var borrower = await _borrowerRepository.AddAsync<CreateBorrowerDto,BorrowerDto>(newBorrower);
-            return CreatedAtAction("GetBorrower", new { id = borrower.Id }, borrower);
+          if (newBorrower == null)
+          {
+                return BadRequest();
+          }
+            var formattedBorrower = _mapper.Map<Borrower>(newBorrower);
+            await _borrowerRepository.AddAsync(formattedBorrower);
+
+
+            return CreatedAtAction("GetBorrower", new { id = formattedBorrower.Id }, formattedBorrower);
         }
 
         // DELETE: api/Borrowers/5
@@ -108,7 +121,11 @@ namespace Lendr.API.Controllers
         [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> DeleteBorrower(int id)
         {
-
+            var borrower = await _borrowerRepository.GetAsync(id);
+            if (borrower == null)
+            {
+                return NotFound();
+            }
             await _borrowerRepository.DeleteAsync(id);
 
             return NoContent();
